# backend/public/api/openapi.yaml
openapi: '3.0.1'
info:
  title: Cadastro de Consultas medicas
  description: 
    CRUD de Consultas para o projeto SimplicaMed que ira cadastrar as consultas dos pacientes, eles devem informar seu nome, sua idade, seu peso e a data da consulta medica
  version: '1.0'
servers:
  - url: http://{host}:{porta}/api
    variables:
      host: 
        default: localhost
      porta:
        default: '8000'
paths:
...
  /profissionals:
    get:
      summary: Acessar profissionais
      responses:
        200:
          description: OK
        404:
          description: Não encontrado.
    post:
      summary: Criar Profissional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nome do profissional
                  example: Juliano
                cpf:
                  type: string
                  description: CPF do profissional
                  example: 123.123.123-09
                endereco:
                  type: string
                  description: Endereco do profissional
                  example: Rua da felicidade, 990
                telefone:
                  type: string
                  description: telefone do profissional
                  example: '(84)99876-1234'
                tipo:
                  type: string
                  description: Tipo de profissional
                  example: Médico cardiologista
      responses:
        201:
          description: Profissional criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                    description: O URI do novo recurso criado.
        404:
          description: Requisição inválida
...